Support projections with zero columns
An example of a SELECT statements without a projection:

```sql
SELECT FROM users;
```

For the curious, the above executes successfully in `psql` but the projection has no columns and no rows are returned. In psql it will report the row count (but that is not part of the result set).

It's apparently used by some DB adapters (we're looking at you `postgrex`) to assert that particular tables exist - the query will fail if it doesn't.

Statements like the above will now type-check.

Additionally, `TypeChecker::statement_type(&self)` now returns
`Some(Projection::Empty)` instead of `None` for statements that can return
a projection.