Attempt to connect and send data to port
Previous method of checking if the port was open (`nc -z`) produced
false-positives, due to how port forwarding works with Docker.

The port would show as being open, even if the service at the other end of the
port forward (inside the container) was actually closed.

This now tries to connect and send data, which should eliminate false-positives.