Introduce TypeCell
Previously after two types had been unified they did not point to the
same allocation in memory: they would each become a copy of the
resulting unified type.

This was problematic because the relationship between node types was not
preserved.

For example, given two AST nodes A and B and their initial
(pre-unification) types TA and TB. Unification computes a new type
TU and TA & TB are in-place mutated to become TU.

Later, type TA is mutated via unification with another node's type but
this is not reflected in TB because TA and TB do not share an
allocation.

TypeCell fixes this.

A TypeCell is initialised with a Type, like this:
TypeCell::new(some_type);

Initially it does not share its underlying allocation with any other
TypeCell.

Given two TypeCells, a and b, `a.join(&b)` permanently ensures that the
underlying allocation of the two TypeCells is shared.

(Under the hood, the `Type` allocation of `a` is dropped and replaced
with a clone of an `Rc<RefCell<..>>` from `b`).

An update to type `Type` stored in `a` is equivalent to updating the
`Type` stored in `b`.