Fix: explicit columns is optional in INSERT
When there were no named columns the source would be type-checked
against an empty projection.

Now, when the columns are not explicity named in an INSERT STATEMENT
then the source (e.g. the VALUES part of the statement) will be
type-checked against the projection that represents all of the columns
from the table.