Resolve DEFAULT keyword in expression as fresh type var
The following is valid SQL in Postgres.

```sql
INSERT INTO some_table (foo, bar) VALUES (123, DEFAULT);
```

The problem was that sqlparser parses `DEFAULT` as an `Expr::Identifier`
which means that we would attempt to find a projection column named
`DEFAULT` in the current scope.

We now check for `DEFAULT` and resolve the expression to a fresh type
variable (which in the example above would eventually be unified with the
type of the `bar` column).

This is *probably* a bug in sqlparser but the workaround is
straightforward.