Fix a weird interaction between binary output and Mise
`mise run test` was consistently failing for me with the following:

```
[test:wait_for_postgres_to_quack] Waiting for localhost:6432
[test:wait_for_postgres_to_quack] panic in task: panicked at src/cmd.rs:312:41:
[test:integration] called `Result::unwrap()` on an `Err` value: Error { kind: InvalidData, message: "stream did not contain valid UTF-8" }
```

For posterity, this was the response output (via od -c) that mise
appeared to be choking on:

```
0000000    R  \f 005   . 372 362 310  \n
0000010
```

This is reliably triggered by the `echo "response: ${response}"` line in
the `if [ -n "${DEBUG}" ]` true branch, but also *sometimes* by the
response echo from nc.

This fix belts-and-suspenders this problem by removing both the echo *and*
condemning nc's stderr to the void.

(This also cleans up some extraneous "-n"s in the echo output; see
https://cipherstash.slack.com/archives/C05SN4WRAAH/p1736736780831759
if you are very curious.)