Resolve unresolved type vars of literals and param placeholders
Prior to this change all of the following statements would have failed to
type check with something like:

`Type(Incomplete("type Var(0) contains unresolved type variables"))`

 - `select 1;`
 - `select t from (select 1 as t);`
 - `select * from (select 1);`
 - `select $1;`
 - `select t from (select $1 as t);`
 - `select * from (select $1);`

This was because:

1. None of the values came from a table-column which would have
   constrained their types.

2. None of the values have been used in an expression that would have
   constrained their types (beyond simple propogation of a projection)

Now, all of the above will resolve successfully by applying the
following typing rule:

When resolving a type variable and no substitution has been found, if
(and only if) that type variable is associated with one or more
`Expr::Value(_)` AST nodes (which represent literals or parameter
usages), then resolve the type as Native.

An alternative solution that would have "worked" would be to
insdiscriminately resolve all unresolved type variables to Native but
that would mask bugs in the unifier where we missed something that
should have resolved to a type other than Native.