Add missing bounds check to `maybe_json` function
This change adds a missing bounds check in the `maybe_json` function.

This function is called on each result returned by a query. The missing bounds
check caused queries with empty results (e.g. empty strings) to panic and
close the connection.

This change fixes the bug by adding in the bounds check.