Replace `expect("ok")` with `unwrap()` in Rust tests
This change replaces calls of `expect("ok")` with `unwrap()` in Rust
tests. Both `expect` and `unwrap` are perfectly fine to use, but the
error messages from `expect` can be noisy if they don't give
meaningful context.

We could update the messages to describe why we expect results
to be `Ok`, but `unwrap`ing is also reasonable.

See the docs on `expect` for more details:
https://doc.rust-lang.org/std/result/enum.Result.html#method.expect