Fix handling of repeated literals in statement transformation
This change fixes a bug where statements containing repeated
literals would fail when at least one of the literals was for an EQL
column.

Previously, literals and their replacements from encryption were
passed to the transformer as a `HashMap<&'ast ast::Expr, ast::Expr>`
(where the key, `&'ast ast::Expr`, represents the original node to
be replaced). Using `&'ast ast::Expr` as a key hashes the _contents_,
of the node. This causes issues when literals are repeated because
the duplicate literals have the same hash.

For example, given the following statement...
```sql
-- Assume both columns are integer columns and that `encrypted_int`
--  is an EQL column.
INSERT INTO encrypted (id, encrypted_int) VALUES (123, 123)
```
... the transformer would incorrectly replace the first node containing
`123` with the ciphertext that was meant for the second node.
This would be surfaced as a type error from Postgres.

To fix this, we need to make sure that hash keys uniquely identify
the node in the original AST. Conveniently, the type checker already
uses `NodeKey` for this purpose, so the fix is to use `NodeKey`s as
hash keys during transformation, too.