# This is a combination of 5 commits. # This is the 1st commit message:
Initial commit (import from sql-mapper)

# This is the commit message #2:

First cut of public API

Call `eql_mapper::type_check(..)`

If successful, that returns a `TypedStatement` which describes:

1. the types & values of any literals that are destined to become EQL
   types and require encrypting
2. all of the expected params and their types
3. the type of the statement's projection (which will inform row
   decryption)

It is up to the caller to check the expected number of params is
correct, and arrange for all literals to be encrypted. Then call:

`TypedStatement::transform(encrypted_literals)`

which will return a new `Statement` with replaced literals.

# This is the commit message #3:

Make TypedStatement::literals a public field

# This is the commit message #4:

Ensure the actual types representing, um.. types are exported

# This is the commit message #5:

Add check to see if statement requires a transformation