First cut of public API (import from sql-mapper)
Call `eql_mapper::type_check(..)`

If successful, that returns a `TypedStatement` which describes:

1. the types & values of any literals that are destined to become EQL
   types and require encrypting
2. all of the expected params and their types
3. the type of the statement's projection (which will inform row
   decryption)

It is up to the caller to check the expected number of params is
correct, and arrange for all literals to be encrypted. Then call:

`TypedStatement::transform(encrypted_literals)`

which will return a new `Statement` with replaced literals.

Also:

- Make TypedStatement::literals a public field
- Ensure the actual types representing, um.. types are exported
- Add check to see if statement requires a transformation
- Export simplified types