feat: warn if a column looks encrypted with no config
In RowDescription, types are returned but it seems like we can only
see JSONB type, not the domain (cs_encrypted_v1). Because of this,
we attempt parsing into an encrypted column, and warn if successful.

No warning will be given if the column has encrypt config or the
Proxy config explicitly disables mapping.