feat: track all `Expr::Value` nodes
`ValueTracker` keeps a reference to all of the `Expr::Value` types
encountered during AST traversal and provides an API that other visitor
pipeline states can interrogate.

It is used to implement the automatic resolution of projection columns
whose type would be unresolved if their type was unused in the
statement.

Unresolved type variables in projection columns can safely be resolved to
`Native` if there is an `Expr::Value` node in the AST with the same type
variable.