Fix: broken tests
A function which is not a known EQL function will always be typed as
returning `NativeValue(None)`.

There is simply no way for `TypeInferencer` to know whether a function
returns the same native type as its argument.

The test previously worked because the `InferType` impl for `Function`
was calling `self.unify_nodes(function, expr)` (where `expr` was the
argument to the function).

I'd noticed that it was incorrect but the test was not correspondingly
updated.