feat: SQL transformations related to `GROUP BY`
Includes a version bump to the latest `sqltk` version which vendors
`sqlparser` as `sqltk-parser` and includes an extra parameter in
`Transform::transform` to allow contextual transformations (instead of
solely by type).

The following refactorings were done as part of this PR:

1. Break sole `Transform` impl into modular rules using the
   `TransformationRule` trait.

2. Removed `EqlFunctionTracker` - this existed to support `Expr`
   transformations in `ORDER BY` clauses. With contextual transformation
   it was no longer required.

3. Removed `TypeCell` - it did not carry its weight and made debugging
   really hard. The `TypeRegistry` now associates nodes to Arc<Type>
   using an intermediary `TypeVar`.

An extra method was added to the post type checking step in the
`EqlMapper` to resolve all `Value` nodes which have no concrete type
after the inference step to `NativeValue`.