feat: tracing of `InferType` impls & `Unifier::unify`
Debugging type checking code in VSCode is suboptimal - we really want a
graphical representation of inference at every step (but also a debugger
that doesn't lock up at inconvenient times would be great).

To make instrumentation consistent I implemented a `#[trace_infer]`
attribute macro which can be used on `InferType` impls.

The `Unifier` has a single `#[tracing::instrument!(..)]` call and there
is also code in `EqlMapper` to dump nodes types and type substitutions
at the end of the type checking process which is useful when debugging
failing inference tests.

To get a trace from a test, uncomment the `init_tracing();` line.

The output is *very* verbose but very useful.