feat: add new `Type::Associated` variant & EqlTraits
The type system requires associated types to represent EQL terms used
only with particular operators and functions. The kinds of encrypted AST
literals (`Value` nodes) values that can be communicated as a result of
an `eql-mapper` type check will be used by Proxy to inform *how* those
terms are to be encrypted.

This commit also introduces `EqlTrait`. In `eql-mapper` `EqlTrait` is
Rust enum but each variant respresents the name of a notional trait that
EQL types can implement.

These are the currently defined EQL traits: `Eq`, `Ord`, `TokenMatch`,
`JsonLike` & `Contain`.

Associated types belong to an `EqlTrait` but are implemented on types
that implement that trait (just like in Rust).

For example the `EqlTrait::JsonLike` has `Accessor` and `Path`
associated types.

An `Accessor` is the type of the expression the right hand side of `->`
or `->>`.

A `Path` is the type of the second argument to `jsonb_query_path`.

An associated type is fundamentally related to its parent type in that
it shares configuration - but cannot be used in a position where the
parent type is mandated.

Lastly, `Type::Var` now has "type bounds". Type bounds are represented
as an `EqlTraits` value (not the pluralisation) which is just a set of
boolean flags: one per `EqlTrait`.