feat: declarative type environment
The current implementation of function type checking in `eql-mapper`
manually instantiates and unifies types for a handful of specific SQL
functions (`MIN`, `MAX`, `COUNT`) and operators (`=`, `<`, `<=`, `>=`,
`>`).

The list of functions and operators is about to grow significantly and
implementing type checking by hand for all of those by manually
instantiating and unifying type variables is going to be tedious and
error prone.

This commit introduces `TypeDecl`, `TypeEnv` and `InstantiatedTypeEnv`.

A `TypeDecl` is a type declaration for the `eql-mapper` type system. A
`TypeDecl` is purely symbollic and cannot be used by the `Unifier`.

What a `TypeDecl` does is achieve a consistent notation for declaring
types and their bounds.

Multiple type declarations can be put into a `TypeEnv` - a `TypeEnv` is
a scope for some related type declarations. Type declarations can
reference other type declarations in the same type env via type
variables `TVar`s.

A `TypeEnv` can be *instantiated*, returning an `InstantiatedTypeEnv`.
In an `InstantiatedTypeEnv` all of the `TypeDecls` have been converted
into `Arc<Type>` values ready to be used by the `Unifier`.

Additionally, the `InstantiatedTypeEnv` allows retrieval of its types
via a `TVar` - it maintains the same associated of `TVar` to type as the
original `TypeEnv`.