feat: support bounded type variables and associated types in `Unifier`
An associated type represents "deferred" unification.

This is a bit more involved than unifying other types. The reason is
that the `impl_ty` (parent type) of the associated type can be a type
variable bounded by the appropriate `EqlTrait` that declares the
associated type.

At the time the `Type::Associated(AssociateType)` is instantiated by the
`TypeInferencer` we cannot guarantee that the parent type has been
resolved - so we can resolve the conrete associated type.

So you can think of the associated type as an "obligation" (that's the
actual term used in type theory).

Whenever unification tries to unify a type with an associated type
obligation, it first checks if the parent type of the associated type
has been resolved and then resolves the associated type.

If the parent type has not been resolved, then the *unresolved*
associated type is unified with the other type as a new
`Type::Associated`, deferring resolution until later.

`Type::Var` now also supports bounds (as an `EqlTraits` struct). Two
type variables always unify: even they have different bounds. The
result is a new type variable with merged bounds.

Whenever a concrete type is unified with a type variable, the effective
bounds of the concrete type MUST contain all of the bounds of
the type variable (concrete types always have *implied* effective
bounds). Thr concrete type is allowed to define additional bounds in
excess of those required by the type variable.