fix: pin containers for dev and test to Postgres 17
Previously we were using images tagged `postgres:latest`, which was
pointing to the latest release in the Postgres 17 series.

Postgres 18 was released in the last 24 hours, and `postgres:latest` now
points to Postgres 18.

There seem to be some changes in the Postgres 18 containers, and volume
mounting now breaks consistently in CI and locally.

For now we will pin to Postgres 17, and look at supporting Postgres 18
later.